+++ CpE76_Tackticus/Unit.h
@@ -11,13 +11,13 @@
 @interface Unit : NSObject
 
 @property NSString *type;
-@property int baseHP;
-@property int baseMP;
+@property float baseHP;
+@property float baseMP;
 @property int baseDamage;
 @property int skillDamage;
 @property int moveRange;
 @property int attackRange;
 @property int skillRange;
 
-- (id)initType:(NSString *)t hp:(int)bh mp:(int)bm bDamage:(int)bd sDamage:(int)sd moveRange:(int)mr atkRange:(int)ar skillRange:(int)sr;
+- (id)initType:(NSString *)t hp:(float)bh mp:(float)bm bDamage:(int)bd sDamage:(int)sd moveRange:(int)mr atkRange:(int)ar skillRange:(int)sr;
 @end
